requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
coverage==7.6.1
crashtest==0.4.1
cryptography==42.0.8
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.5
fastjsonschema==2.20.0
filelock==3.15.4
google-auth==2.33.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
importlib_metadata==8.0.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==24.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg==3.2.1
psycopg-binary==3.2.1
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2c
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
pyproject_hooks==1.1.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.19
starlette==0.37.2
tomli==2.0.1
tomlkit==0.12.5
trove-classifiers==2024.7.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
zipp==3.19.2
loguru==0.7.3
